Some refactoring in tracker
