+++ clomonitor-tracker/Cargo.toml
@@ -8,6 +8,7 @@ rust-version = "1.58"
 
 [dependencies]
 anyhow = "1.0.64"
+async-trait = "0.1.57"
 clap = { version = "3.2.20", features = ["derive"] }
 clomonitor-core = { path = "../clomonitor-core" }
 config = "0.13.2"
